Task_28 decimal

+++ Task_28/Program.cs
@@ -5,11 +5,11 @@
 // 5 -> 120
 Console.Write("Введите любое число: ");
 int enteredNumber = int.Parse(Console.ReadLine());
-int result = Fractal(enteredNumber);
+decimal result = Fractal(enteredNumber);
 Console.WriteLine($"Вычисленное значение от введенного числа {enteredNumber} по условию равно {result}");
 
-int Fractal(int number){
-    int answer = 1;
+decimal Fractal(int number){
+    decimal answer = 1;
     for(int i = 1; i <= number; i++){
         try{
             checked
